move the force resync button
