+++ front/components/workspace/WorkOSConnection.tsx
@@ -22,8 +22,10 @@ import {
   useCreateWorkOSOrganization,
   useWorkOSAdminPortalUrl,
 } from "@app/lib/swr/workos";
+import { useSyncWorkOSDirectoriesAndUsers } from "@app/lib/swr/workspaces";
 import { WorkOSPortalIntent } from "@app/lib/types/workos";
 import logger from "@app/logger/logger";
+import type { WorkspaceType } from "@app/types";
 
 const ADMIN_PANEL_OPTIONS = {
   domain: [
@@ -62,11 +64,42 @@ const ADMIN_PANEL_OPTIONS = {
   ],
 };
 
-interface WorkOSConnectionProps {
-  owner: {
-    sId: string;
-    workOSOrganizationId?: string | null;
+interface WorkOSSyncButtonProps {
+  owner: WorkspaceType;
+}
+
+export function WorkOSSyncButton({ owner }: WorkOSSyncButtonProps) {
+  const [isLoading, setIsLoading] = useState(false);
+  const { triggerFullSync } = useSyncWorkOSDirectoriesAndUsers(owner);
+
+  const handleSync = async () => {
+    setIsLoading(true);
+    await triggerFullSync();
+    setIsLoading(false);
   };
+
+  return owner.workOSOrganizationId ? (
+    <Button
+      variant="primary"
+      onClick={handleSync}
+      disabled={isLoading}
+      label={isLoading ? "Syncing..." : "Sync WorkOS Directories & Groups"}
+    />
+  ) : (
+    <Button
+      variant="primary"
+      disabled
+      label={
+        isLoading
+          ? "Syncing..."
+          : "Your workspace is not linked to a WorkOS organization."
+      }
+    />
+  );
+}
+
+interface WorkOSConnectionProps {
+  owner: WorkspaceType;
 }
 
 export function WorkOSConnection({ owner }: WorkOSConnectionProps) {
@@ -214,6 +247,11 @@ export function WorkOSConnection({ owner }: WorkOSConnectionProps) {
           />
         </DialogContent>
       </Dialog>
+      {/* Debug button: will be replaced by an actual admin console */}
+      <Page.P variant="secondary">Synchronize your directories.</Page.P>
+      <div className="flex w-full flex-col items-start gap-3">
+        <WorkOSSyncButton owner={owner} />
+      </div>
     </Page.Vertical>
   );
 }
